+++ src/main/java/ladder/domain/BooleanGenerator.java
@@ -1,0 +1,6 @@
+package ladder.domain;
+
+public interface BooleanGenerator {
+
+    boolean generate();
+}
