+++ src/main/java/ladder/domain/Ladder.java
@@ -1,11 +1,11 @@
 package ladder.domain;
 
-import static java.util.stream.Collectors.*;
+import static java.util.stream.Collectors.collectingAndThen;
+import static java.util.stream.Collectors.toList;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
-import ladder.utils.RandomGenerator;
 
 public class Ladder {
     private final List<Line> lines;
@@ -14,8 +14,8 @@ public class Ladder {
         this.lines = lines;
     }
 
-    public static Ladder generate(int countOfLine, int countOfBar, RandomGenerator randomGenerator) {
-        return Stream.generate(() -> Line.generate(countOfBar, randomGenerator))
+    public static Ladder generate(int countOfLine, int countOfBar, BooleanGenerator booleanGenerator) {
+        return Stream.generate(() -> Line.generate(countOfBar, booleanGenerator))
                 .limit(countOfLine)
                 .collect(collectingAndThen(toList(), Ladder::new));
     }
