--- src/main/java/ladder/utils/RandomGenerator.java
@@ -1,6 +1,0 @@
-package ladder.utils;
-
-public interface RandomGenerator<T> {
-
-    T generate();
-}
