+++ src/test/java/ladder/domain/LadderTest.java
@@ -1,8 +1,7 @@
 package ladder.domain;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
-import ladder.utils.BooleanGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -13,7 +12,7 @@ class LadderTest {
     void shouldHasSizeOfInputWhenCreateLadder() {
         // given
         // when
-        Ladder ladder = Ladder.generate(3, 4, new BooleanGenerator());
+        Ladder ladder = Ladder.generate(3, 4, new RandomBooleanGenerator());
         //then
         assertThat(ladder.getLines()).hasSize(3);
     }
